Página web(Plantilla Bootstrap 4) 21 de febrero 2020 Camilo Colón

+++ appsiel/app/Http/Controllers/web/PaginaController.php
@@ -305,6 +305,12 @@ class PaginaController extends Controller
     public function destroy($id){
 
        $pagina = Pagina::find($id);
+       if($pagina->widgets->count() > 0){
+           return response()->json([
+               'status' => 'error',
+               'message' => 'la pagina selecionada tiene secciones agregadas'
+           ]);
+       }
 
        $flag =  $pagina->delete();
